docs/workshop: Use containers instead of VMs for IPA workshop
Using containers instead of VMs for the workshop reduces the
installation of extra software to run the workshop, and makes it easier
to run the workshop on other platforms such as macOS and Windows.

This patch removes the dependency on Vagrant and uses ipalab-config
(http://github.com/rjeffman/ipalab-config) to create a container compose
configuration that can be used with 'podman-compose' to create a cluster
environment to run the workshop.

The workshop has been tested under Linux and Apple's macOS, but should
also run under Microsoft Windows.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>